'Twas nonconformant therefore we fixed it posthaste
We added license information to ensure this script respects thy freedoms.
We also made sure that  it works on Chromium giving thee the option
of choosing they favourite browser (as long as it's either
Firefox, Chromium or similar).
Finally we decided to rest our vocal chords and not to commit
using our royal canterlot voice.